@@ -4,7 +4,7 @@ version: 0.2.1
 homepage: https://github.com/deansyd/icloud_storage
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.20.0"
 
 dependencies:
